feat: div instruction (and virtual const)

+++ zkml-jolt-core/src/utils/mod.rs
@@ -1,4 +1,4 @@
 /// Helper function to convert Vec<u64> to iterator of i128
 pub fn u64_vec_to_i128_iter(vec: &[u64]) -> impl Iterator<Item = i128> + '_ {
-    vec.iter().map(|v| *v as i128)
+    vec.iter().map(|v| *v as u32 as i32 as i64 as i128)
 }
